requirements.txt
# TikTok Video Analysis System - Requirements
# GPU-optimized ML video analysis with 17 active analyzers
# Python 3.10+ required for optimal compatibility

# Core ML Frameworks
torch>=2.0.0,<2.5.0          # PyTorch with CUDA support
torchvision>=0.15.0,<0.20.0  # Computer vision operations
torchaudio>=2.0.0,<2.5.0     # Audio processing
transformers>=4.30.0,<4.45.0 # HuggingFace transformers
accelerate>=0.20.0,<0.35.0   # Model acceleration

# GPU Optimization
nvidia-ml-py3>=7.352.0        # NVIDIA GPU monitoring
pynvml>=11.0.0               # NVIDIA management library

# Video Processing
opencv-python>=4.8.0,<4.10.0  # Computer vision
opencv-contrib-python>=4.8.0,<4.10.0  # Additional OpenCV modules
moviepy>=1.0.3,<1.1.0        # Video editing
imageio>=2.31.0,<2.35.0      # Image I/O
imageio-ffmpeg>=0.4.8,<0.5.0 # FFmpeg backend
av>=10.0.0,<12.0.0           # PyAV for video processing

# Video Download
yt-dlp>=2023.7.6             # YouTube/TikTok downloader
requests>=2.31.0,<2.32.0    # HTTP requests
urllib3>=1.26.0,<2.0.0      # URL handling

# Computer Vision Models
ultralytics>=8.0.0,<8.3.0   # YOLOv8 for object detection
mediapipe>=0.10.0,<0.11.0   # Google MediaPipe
insightface>=0.7.3,<0.8.0   # Face recognition

# Text Recognition and Processing
easyocr>=1.7.0,<1.8.0       # Optical Character Recognition
pytesseract>=0.3.10,<0.4.0  # Tesseract OCR wrapper
Pillow>=9.5.0,<11.0.0       # Image processing
pdf2image>=1.16.0,<1.17.0   # PDF to image conversion

# Audio Processing
librosa>=0.10.0,<0.11.0     # Audio analysis
soundfile>=0.12.0,<0.13.0   # Audio file I/O
webrtcvad>=2.0.10,<2.1.0    # Voice activity detection
pyaudio>=0.2.11,<0.3.0      # Audio I/O
pydub>=0.25.1,<0.26.0       # Audio manipulation

# Speech Recognition
openai-whisper>=20230314,<20240101  # OpenAI Whisper
faster-whisper>=0.8.0,<0.9.0       # Optimized Whisper

# Natural Language Processing
spacy>=3.6.0,<3.8.0         # NLP library
nltk>=3.8.1,<3.9.0          # Natural Language Toolkit
textblob>=0.17.1,<0.18.0    # Text processing
sentence-transformers>=2.2.0,<2.3.0  # Sentence embeddings

# Machine Learning Utilities
scikit-learn>=1.3.0,<1.4.0  # ML algorithms
scipy>=1.10.0,<1.12.0       # Scientific computing
numpy>=1.24.0,<1.26.0       # Numerical computing
pandas>=2.0.0,<2.2.0        # Data manipulation
matplotlib>=3.7.0,<3.8.0    # Plotting
seaborn>=0.12.0,<0.14.0     # Statistical visualization

# Image Processing
Pillow>=9.5.0,<11.0.0       # PIL fork
skimage>=0.21.0,<0.22.0     # Scikit-image
plotly>=5.15.0,<5.18.0      # Interactive plotting

# Web Framework
fastapi>=0.100.0,<0.105.0   # Modern web framework
uvicorn>=0.22.0,<0.25.0     # ASGI server
pydantic>=2.0.0,<2.6.0      # Data validation
python-multipart>=0.0.6,<0.1.0  # Form data parsing
jinja2>=3.1.0,<3.2.0        # Template engine

# Database and Storage
sqlalchemy>=2.0.0,<2.1.0    # Database ORM
alembic>=1.11.0,<1.13.0     # Database migrations
psycopg2-binary>=2.9.6,<2.10.0  # PostgreSQL adapter
redis>=4.5.0,<5.1.0         # Redis client

# Async Processing
celery>=5.3.0,<5.4.0        # Distributed task queue
flower>=2.0.0,<2.1.0        # Celery monitoring
kombu>=5.3.0,<5.4.0         # Messaging library

# Configuration and Environment
python-dotenv>=1.0.0,<1.1.0 # Environment variables
pyyaml>=6.0,<6.1             # YAML parsing
toml>=0.10.2,<0.11.0        # TOML parsing
configparser>=5.3.0,<5.4.0  # Configuration files

# Monitoring and Logging
psutil>=5.9.0,<5.10.0       # System monitoring
GPUtil>=1.4.0,<1.5.0        # GPU utilities
py-cpuinfo>=9.0.0,<9.1.0    # CPU information
prometheus-client>=0.17.0,<0.18.0  # Metrics collection

# Development and Testing
pytest>=7.4.0,<7.5.0        # Testing framework
pytest-asyncio>=0.21.0,<0.22.0  # Async testing
black>=23.0.0,<24.0.0       # Code formatting
flake8>=6.0.0,<6.1.0        # Linting
isort>=5.12.0,<5.13.0       # Import sorting
mypy>=1.4.0,<1.6.0          # Type checking

# Data Serialization
msgpack>=1.0.5,<1.1.0       # Binary serialization
orjson>=3.9.0,<3.10.0       # Fast JSON library
pickle5>=0.0.12,<0.1.0      # Pickle protocol 5
joblib>=1.3.0,<1.4.0        # Lightweight pipelining

# Utility Libraries
tqdm>=4.65.0,<4.67.0        # Progress bars
click>=8.1.0,<8.2.0         # Command line interface
rich>=13.4.0,<13.6.0        # Rich text and beautiful formatting
colorama>=0.4.6,<0.5.0      # Cross-platform colored terminal
python-dateutil>=2.8.2,<2.9.0  # Date utilities

# Performance Profiling
memory-profiler>=0.60.0,<0.61.0  # Memory usage profiling
line-profiler>=4.0.0,<4.2.0      # Line-by-line profiling
py-spy>=0.3.14,<0.4.0            # Sampling profiler

# HTTP and API clients
httpx>=0.24.0,<0.26.0       # Async HTTP client
aiohttp>=3.8.0,<3.9.0      # Async HTTP client/server
websockets>=11.0.0,<12.0.0 # WebSocket support

# Specialized ML Models
clip-by-openai>=1.0         # CLIP model
sentence-transformers>=2.2.0 # Sentence embeddings
face-recognition>=1.3.0,<1.4.0  # Face recognition
fer>=22.4.0,<23.0.0         # Facial emotion recognition

# Time Series and Analysis
statsmodels>=0.14.0,<0.15.0 # Statistical models
networkx>=3.1,<3.2          # Network analysis

# System Dependencies (install via apt/yum)
# - ffmpeg
# - libsm6 
# - libxext6
# - libxrender-dev
# - libgomp1
# - libglib2.0-0
# - libgl1-mesa-glx
# - libglib2.0-0

# GPU Drivers (install separately)
# - nvidia-driver-535 or newer
# - cuda-toolkit-12-4 or newer
# - libcudnn8-dev

# Optional: Cloud Storage Integration
boto3>=1.28.0,<1.30.0       # AWS SDK
google-cloud-storage>=2.10.0,<2.11.0  # Google Cloud Storage
azure-storage-blob>=12.17.0,<12.18.0  # Azure Blob Storage

# Optional: Advanced Analytics
wandb>=0.15.0,<0.16.0       # Experiment tracking
tensorboard>=2.13.0,<2.15.0 # TensorBoard logging
mlflow>=2.5.0,<2.7.0       # ML lifecycle management

# Production Deployment
gunicorn>=21.2.0,<21.3.0    # WSGI HTTP Server
supervisor>=4.2.5,<4.3.0    # Process control system